-hook personalizado para validar los datos de la sesión
